presentation/src/test/java/com/alexmumo/presentation/search/SearchScreenTest.kt
package com.alexmumo.presentation.search

import androidx.compose.ui.test.assertIsDisplayed
import androidx.compose.ui.test.junit4.createComposeRule
import androidx.compose.ui.test.onNodeWithTag
import androidx.navigation.compose.rememberNavController
import org.junit.Before
import org.junit.Rule
import org.junit.Test
import org.junit.runner.RunWith
import org.robolectric.RobolectricTestRunner
import org.robolectric.shadows.ShadowLog

@RunWith(RobolectricTestRunner::class)
class SearchScreenTest {

    @get:Rule
    val composeRule = createComposeRule()

    @Before
    fun setUp() {
        ShadowLog.stream = System.out
    }

    @Test
    fun `verify search screen is displayed`() {

        composeRule.setContent {
            SearchScreen(navController = )
        }
        composeRule.onNodeWithTag("search_tag").assertExists()
        composeRule.onNodeWithTag("search_tag").assertIsDisplayed()
    }
}